wine_pairing/requirements.txt
langchain
langchain_openai
ipython
streamlit
python-dotenv
tabulate